requirements.txt
# ====================
# Nutrition AI - Production Dependencies
# Version: 2.0.0
# ====================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# AI Model - Google Gemini (Primary)
google-genai>=1.46.0

# Data Validation & Models
pydantic>=2.12.0

# HTTP & Async Requests
aiohttp==3.9.1
requests==2.32.0

# Image Processing (Barcode Detection)
Pillow==10.1.0
pyzbar==0.1.9
opencv-python==4.8.1.78
numpy<2

# Environment Configuration
python-dotenv==1.0.0

# Note: anthropic removed as we now use Google Gemini exclusively
# Database: Using built-in sqlite3 (no additional package needed)
